[1780] Restore compile after changes in KilledMonsterCredit function. Requires Mangos 10363
Restore compile after changes in KilledMonsterCredit function. Requires Mangos 10363

git-svn-id: https://scriptdev2.svn.sourceforge.net/svnroot/scriptdev2@1780 5f9c896b-1e26-0410-94da-f77f675e2462

+++ scripts/eastern_kingdoms/karazhan/boss_prince_malchezaar.cpp
@@ -100,7 +100,7 @@ struct MANGOS_DLL_DECL netherspite_infernalAI : public ScriptedAI
 
     uint32 HellfireTimer;
     uint32 CleanupTimer;
-    uint32 malchezaar;
+    uint64 malchezaar;
     InfernalPoint *point;
 
     void Reset() {}
